spi: behave better when CS is not provided
Instead of assuming that the initial state of the CLK line is 0, assume
that it's the same as the first sample we have of that line.

This makes no difference for decoding when the chip select line is
provided, as that resets the state machines.

When CS is not provided however, this makes decodes actually work for
spi modes other than 0,0.

Signed-off-by: Karl Palsson <karlp@etactica.com>